Calcular porcentagem salário

+++ src/CalculadoraImposto.java
@@ -9,7 +9,7 @@ public class CalculadoraImposto {
     public static void main(String[] args){
         float salario = 2500.00F;
         float porcentagem = 30L;
-        float porcentagemDoSalario = salario * (porcentagem / 100);
+        double porcentagemDoSalario = salario * (porcentagem / 100);
         System.out.println(porcentagemDoSalario);
     }
 }
